remove NewFromPool, clarify Close behavior, improve NewFromConn tests
NewFromPool was flawed because the Driver model here is to call Close()
on the driver at the end of the migration, which was closing the pool. I
could have tried to fix this to avoid doing that when running against a
pool, however I decided it makes more sense to trim down the interface
here and only offer two constructors: New and NewFromConn. Logic is
almost fully reused between the two.

With New, any created conn is closed when the Driver is closed (because
it is considered "owned" by the driver). With NewFromConn, the provided
conn is left alone even upon driver.Close().